# Lokata.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lokata.WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Place/{id}": {
      "get": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/TakePlace": {
      "get": {
        "tags": [
          "TakePlace"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TakePlace"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TakePlace"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TakePlace"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TakePlace"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TakePlace"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TakePlace"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TakePlace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TakePlace/{id}": {
      "get": {
        "tags": [
          "TakePlace"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TakePlace"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakePlace"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakePlace"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TakePlace"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TakePlace"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TakePlace"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TakePlace"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TakePlace"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TargetsAndCardsPhoto": {
      "get": {
        "tags": [
          "TargetsAndCardsPhoto"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetsAndCardsPhoto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetsAndCardsPhoto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetsAndCardsPhoto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TargetsAndCardsPhoto"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsAndCardsPhoto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsAndCardsPhoto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsAndCardsPhoto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TargetsAndCardsPhoto/{id}": {
      "get": {
        "tags": [
          "TargetsAndCardsPhoto"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsAndCardsPhoto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsAndCardsPhoto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsAndCardsPhoto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TargetsAndCardsPhoto"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsAndCardsPhoto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsAndCardsPhoto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsAndCardsPhoto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TargetsAndCardsPhoto"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TargetsOrCardTake": {
      "get": {
        "tags": [
          "TargetsOrCardTake"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetsOrCardTake"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetsOrCardTake"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetsOrCardTake"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TargetsOrCardTake"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsOrCardTake"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsOrCardTake"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsOrCardTake"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TargetsOrCardTake/{id}": {
      "get": {
        "tags": [
          "TargetsOrCardTake"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsOrCardTake"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsOrCardTake"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsOrCardTake"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TargetsOrCardTake"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsOrCardTake"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsOrCardTake"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TargetsOrCardTake"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TargetsOrCardTake"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "postalPlace": {
            "type": "string",
            "nullable": true
          },
          "places": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Place"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Approach": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competitionsId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "instructorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "umpireId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competition": {
            "$ref": "#/components/schemas/Competition"
          },
          "competitions": {
            "$ref": "#/components/schemas/Competitions"
          },
          "instructor": {
            "$ref": "#/components/schemas/Instructor"
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          },
          "umpire": {
            "$ref": "#/components/schemas/Umpire"
          }
        },
        "additionalProperties": false
      },
...
      "Competition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "seriesCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPointsPerSeries": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Competitions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "placeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "additionalProperties": false
      },
      "Competitor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "dateOfBirth": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDisabledPerson": {
            "type": "boolean",
            "nullable": true
          },
          "professional": {
            "type": "boolean",
            "nullable": true
          },
          "sexId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "shootingPlacesCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "addressNavigation": {
            "$ref": "#/components/schemas/Address"
          },
          "competitionsList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Competitions"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TakePlace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tookPlace": {
            "type": "boolean",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "TargetsAndCardsPhoto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "photo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "TargetsOrCardTake": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cardOrTargetTaken": {
            "type": "boolean",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}